requirement.txt
Package           Version    
----------------- -----------
arandr            0.1.9      
asn1crypto        0.24.0     
automationhat     0.2.0      
blinker           1.4        
blinkt            0.1.2      
buttonshim        0.0.2      
Cap1xxx           0.1.3      
certifi           2018.8.24  
chardet           3.0.4      
Click             7.0        
colorama          0.3.7      
colorzero         1.1        
configparser      3.5.0b2    
cookies           2.2.1      
cryptography      2.6.1      
dnspython         1.16.0     
drumhat           0.1.0      
entrypoints       0.3        
enum34            1.1.6      
envirophat        1.0.0      
ExplorerHAT       0.4.2      
Flask             1.0.2      
fourletterphat    0.1.0      
funcsigs          1.0.2      
gpg               1.12.0     
gpiozero          1.5.1      
idna              2.6        
ipaddress         1.0.17     
itsdangerous      0.24       
Jinja2            2.10       
keyring           17.1.1     
keyrings.alt      3.1.1      
MarkupSafe        1.1.0      
microdotphat      0.2.1      
mock              2.0.0      
mote              0.0.4      
motephat          0.0.3      
numpy             1.16.2     
oauthlib          2.1.0      
olefile           0.46       
pantilthat        0.0.7      
pbr               4.2.0      
phatbeat          0.1.1      
pianohat          0.1.0      
picamera          1.13       
piglow            1.2.5      
pigpio            1.44       
Pillow            5.4.1      
pip               18.1       
pycairo           1.16.2     
pycrypto          2.6.1      
pygame            1.9.4.post1
PyGObject         3.30.4     
pyinotify         0.9.6      
PyJWT             1.7.0      
pyOpenSSL         19.0.0     
pyserial          3.4        
pyxdg             0.25       
rainbowhat        0.1.0      
requests          2.21.0     
requests-oauthlib 1.0.0      
responses         0.9.0      
RPi.GPIO          0.7.0      
RPIO              0.10.0     
RTIMULib          7.2.1      
scrollphat        0.0.7      
scrollphathd      1.2.1      
SecretStorage     2.3.1      
sense-hat         2.2.0      
setuptools        40.8.0     
simplejson        3.16.0     
six               1.12.0     
skywriter         0.0.7      
sn3218            1.2.7      
spidev            3.4        
touchphat         0.0.1      
twython           3.7.0      
unicornhathd      0.0.4      
urllib3           1.24.1     
Werkzeug          0.14.1     
wheel             0.32.3     
